ARUHA-523 Storage service: list, get, create, and delete storages

+++ src/main/java/org/zalando/nakadi/service/StorageService.java
@@ -1,16 +1,18 @@
 package org.zalando.nakadi.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.json.JSONException;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 import org.zalando.nakadi.domain.Storage;
+import org.zalando.nakadi.domain.Timeline;
 import org.zalando.nakadi.repository.db.StorageDbRepository;
+import org.zalando.nakadi.service.timeline.TimelineService;
 import org.zalando.problem.Problem;
 
 import javax.ws.rs.core.Response;
-import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -22,11 +24,15 @@ public class StorageService {
 
     final ObjectMapper objectMapper;
     final StorageDbRepository storageDbRepository;
+    final TimelineService timelineService;
 
     @Autowired
-    public StorageService(final ObjectMapper objectMapper, final StorageDbRepository storageDbRepository) {
+    public StorageService(final ObjectMapper objectMapper,
+                          final StorageDbRepository storageDbRepository,
+                          final TimelineService timelineService) {
         this.objectMapper = objectMapper;
         this.storageDbRepository = storageDbRepository;
+        this.timelineService = timelineService;
     }
 
     public List<Storage> listStorages() {
@@ -44,28 +50,49 @@ public class StorageService {
     }
 
     public Result<Storage> createStorage(final JSONObject storageDetails) {
-        final String type = storageDetails.getString("storage_type");
-        final String configuration = storageDetails.get("configuration").toString();
+        final String type;
+        final JSONObject configuration;
+        final Storage.KafkaConfiguration kafkaConfig;
+
+        try {
+            type= storageDetails.getString("storage_type");
+            configuration = storageDetails.getJSONObject("configuration");
+            kafkaConfig = parseKafkaConfiguration(configuration);
+        } catch (JSONException e) {
+            return Result.problem(Problem.valueOf(UNPROCESSABLE_ENTITY, e.getMessage()));
+        }
         final Storage storage = new Storage();
         storage.setId(UUID.randomUUID().toString());
         storage.setType(Storage.Type.valueOf(type.toUpperCase()));
-        try {
-            storage.parseConfiguration(objectMapper, configuration);
-            storageDbRepository.createStorage(storage);
-        } catch (IOException e) {
-            return Result.problem(Problem.valueOf(UNPROCESSABLE_ENTITY, e.getMessage()));
-        } catch (DataAccessException e) {
-            return Result.problem(Problem.valueOf(Response.Status.INTERNAL_SERVER_ERROR, e.getMessage()));
-        }
+        storage.setConfiguration(kafkaConfig);
+        storageDbRepository.createStorage(storage);
         return Result.ok(storage);
     }
 
+    private Storage.KafkaConfiguration parseKafkaConfiguration(final JSONObject configuration) {
+        final Storage.KafkaConfiguration config = new Storage.KafkaConfiguration();
+        config.setZkAddress(configuration.getString("zk_address"));
+        config.setZkPath(configuration.getString("zk_path"));
+        return config;
+    }
+
     public Result<Void> deleteStorage(final String id) {
+        if (!storageDbRepository.getStorage(id).isPresent()) {
+            return Result.notFound("No storage with ID " + id);
+        }
+        if (isInUse(id)) {
+            return Result.forbidden("Storage " + id + " is in use");
+        }
         try {
             storageDbRepository.deleteStorage(id);
         } catch (DataAccessException e) {
             return Result.problem(Problem.valueOf(Response.Status.INTERNAL_SERVER_ERROR, e.getMessage()));
         }
         return Result.ok();
+    }
+
+    private boolean isInUse(final String id) {
+        final List<Timeline> timelines = timelineService.listTimelines();
+        return timelines.stream().anyMatch(tl -> tl.getStorage().getId().equals(id));
     }
 }
